Simplify subscriptions by removing StompMessage
This is not necessary anymore as users can add a conversion method on a per-subscription basis.
If they are interested in headers, they can use them by providing a custom conversion function.
Therefore, we don't the extra object allocations for the most common use case where just an object is expected.