Fix race conditions in JDK11 websocket listener
The completion of the returned CompletionStage is only used to reclaim the CharSequence.
The onText() method itself can be called again as soon as it completes, which can cause
concurrency issues. This re-entrance is however controlled by the invocations counter,
in turn controlled the call to request(n).
The call to request(1) located after the message processing thus prevents issues.

Resolves:
https://github.com/joffrey-bion/krossbow/issues/42