Fix heartbeater deadlocks
When part of the heartbeater is not started (either send or receive),
notify calls (to reset the timer of that part) would just hang forever.

This is due to the fact that the coroutine consuming from the resetter
channel is never started, and thus this rendez-vous channel suspends
senders forever when calling reset().

Resolves:
https://github.com/joffrey-bion/krossbow/issues/46