Prevent double cleanup of stomp session
Cleanup of the session may happen on normal closure, or on error frames, or many other reasons.
The try-catch in the websocket listener coroutine allows to trigger cleanup when an exception
is thrown in the websocket-frame-processing code.

The problem is that the websocket listener coroutine is cancelled when cleaning up the session.
Since cancellation is an exception, it triggers the try-catch like other exceptions and
triggers a second cleanup.

This is at best annoying when debugging, but for sure it doesn't make sense for it to happen.
Also, the clean up code will evolve with time and it may become a problem to execute it twice.