Fix a crash in DarwinWebSocketClient when using the old memory model
While the websocket client works perfectly fine when in use in an app that adopted the new experimental memory model, it crashes consistently during the handshake phase with the old memory model.

The crash is related to memory corruption with the `NSURLSessionWebSocketTask`, which makes me think it’s about an object that we should be freezing, or not accessing from the right thread.

I couldn’t figure out how to fix the root cause other than enforcing access to it from the same thread, which this achieves by always using `withContext(Dispatchers.Main)`.

The amount of processing done seems pretty light so it shouldn’t be too problematic to have it happen on the main thread?

This was tested by:
- using this for a few weeks on my own project
- making sure CI passed