[test] Improve custom headers test
So far the test server only sent a frame back in case custom headers were sent.
This leads to weird timeouts waiting for that frame in case we expect custom
headers support to work, but it doesn't work (headers are not sent but the
client also doesn't fail).
Another problem is that the test server needs to distinguish "custom" headers
from "standard" headers to decide if it sends a frame back.

Instead, we introduce a special URL path `/echoHeaders` which triggers this
frame for this specific test, and not in other tests. This means we can always
expect a frame in the test, and the server always sends it with all headers.