[doc] Remove redundant list of config properties from the docs
The config KDoc is the most up-to-date source of truth and is already
published. There is already a link to this KDoc from the config section
of the general documentation.
Keeping a manual list of these configuration options is not nice because
it can (and did) drift and become obsolete.